// src/api/generated/model/catalogProductDetailResponse.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Ecommerce Copa API
 * API backend do Ecommerce Copa com autenticação por sessão, catálogo público e CRUD administrativo.
 * OpenAPI spec version: 1.0.0
 */
import type { CatalogCategoryResponse } from "./catalogCategoryResponse.ts";
import type { CatalogProductDetailResponseSchema } from "./catalogProductDetailResponseSchema.ts";
import type { CatalogSkuOptionResponse } from "./catalogSkuOptionResponse.ts";

export interface CatalogProductDetailResponse {
	id?: string;
	category?: CatalogCategoryResponse;
	/** Schema de seletores do produto. */
	schema?: CatalogProductDetailResponseSchema;
	selectedSkuId?: string;
	selectedSku?: CatalogSkuOptionResponse;
	skus?: CatalogSkuOptionResponse[];
}