// src/api/generated/model/reviewResponse.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Ecommerce Copa API
 * API backend do Ecommerce Copa com autenticação por sessão, catálogo público e CRUD administrativo.
 * OpenAPI spec version: 1.0.0
 */

export interface ReviewResponse {
	clientId?: string;
	clientName?: string;
	skuId?: string;
	skuTitle?: string;
	stars?: number;
	comment?: string;
	createdAt?: string;
}